test: Use deep hypothesis strategy for all inputs
Now that we have a deep generation strategy, we’ll apply it to all
tests.
